+++ cloud-project/src/main/java/it/unipi/hadoop/LetterFrequencyCombiner.java
@@ -3,18 +3,18 @@ import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 
-public class LetterFrequencyCombiner extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class LetterFrequencyCombiner extends Reducer<Text, LongWritable, Text, LongWritable> {
 
     
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
-        for (IntWritable val : values) {
+    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+        long sum = 0;
+        for (LongWritable val : values) {
             sum += val.get();
         }
-        context.write(key, new IntWritable(sum)); // Invio al reducer
+        context.write(key, new LongWritable(sum)); // Invio al reducer
     }
 }
