+++ cloud-project/src/main/java/it/unipi/hadoop/LetterFrequencyMapper.java
@@ -2,14 +2,13 @@ package it.unipi.hadoop;
 import org.apache.hadoop.mapreduce.Mapper;
 import java.io.IOException;
 import java.text.Normalizer;
-import org.apache.hadoop.io.IntWritable;
-import org.apache.hadoop.io.Text;
 import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.Text;
 
 
-public class LetterFrequencyMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+public class LetterFrequencyMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
 
-    private final static IntWritable one = new IntWritable(1);
+    private final static LongWritable one = new LongWritable(1);
     private Text letter = new Text();
 
     @Override
