+++ cloud-project/src/main/java/it/unipi/hadoop/LetterFrequencyReducer.java
@@ -5,14 +5,14 @@ import org.apache.hadoop.io.*;
 import org.apache.hadoop.mapreduce.Reducer;
 import java.io.IOException;
 import java.util.HashMap;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.DoubleWritable;
 
 
-public class LetterFrequencyReducer extends Reducer<Text, IntWritable, Text, DoubleWritable> {
+public class LetterFrequencyReducer extends Reducer<Text, LongWritable, Text, DoubleWritable> {
 
     private long totalLetters = 0; // Contatore totale delle lettere
-    private Map<Text, IntWritable> letterSums = new HashMap<>();
+    private Map<Text, LongWritable> letterSums = new HashMap<>();
 
     @Override
     protected void setup(Context context) throws IOException, InterruptedException {
@@ -23,18 +23,18 @@ public class LetterFrequencyReducer extends Reducer<Text, IntWritable, Text, Dou
     }
 
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
-        for (IntWritable val : values) {
+    public void reduce(Text key, Iterable<LongWritable> values, Context context) throws IOException, InterruptedException {
+        long sum = 0;
+        for (LongWritable val : values) {
             sum += val.get();
         }
-        letterSums.put(new Text(key), new IntWritable(sum));
+        letterSums.put(new Text(key), new LongWritable(sum));
     }
 
     @Override
     public void cleanup(Context context) throws IOException, InterruptedException {
         for(Text letter: letterSums.keySet()) {
-            int sum = letterSums.get(letter).get();
+            long sum = letterSums.get(letter).get();
             double frequency = (double) sum / totalLetters;
             context.write(letter, new DoubleWritable(frequency));
         }
